refactor: 네비게이션바 별도 파일 분리
