+++ NewsFit/NewsFit/AppEssential/SceneDelegate.swift
@@ -16,38 +16,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     guard let scene = (scene as? UIWindowScene) else { return }
     
     window = UIWindow(windowScene: scene)
-    let vc = CustomNavigationController(rootViewController: NewsFitHomeViewController())
+    let vc = NewsFitHomeNavigationController(rootViewController: NewsFitHomeViewController())
     window?.rootViewController = vc
     window?.makeKeyAndVisible()
-  }
-}
-
-fileprivate final class CustomNavigationController: UINavigationController {
-  let logoImageView: UIImageView = {
-    let imageView = UIImageView()
-    imageView.image = .newsFitLogo
-    imageView.contentMode = .scaleAspectFit
-    return imageView
-  }()
-  
-  override func viewDidLoad() {
-    super.viewDidLoad()
-    
-    configure()
-  }
-  
-  func configure() {
-    navigationBar.addSubview(logoImageView)
-    logoImageView.snp.makeConstraints { make in
-      make.leading.equalToSuperview()
-      make.centerY.equalToSuperview()
-      make.height.equalToSuperview().offset(-20)
-    }
-    
-    let appearance = UINavigationBarAppearance()
-    appearance.backgroundColor = .white
-    
-    navigationBar.standardAppearance = appearance
-    navigationBar.scrollEdgeAppearance = appearance
   }
 }
